+++ test/sim/test_sim_buggify.c
@@ -61,7 +61,8 @@ pong(void *arg)
 			atomic_fetch_add_explicit(&g_delivered, 1,
 			    memory_order_relaxed);
 			long h = atomic_load_explicit(&g_hash, memory_order_relaxed);
-			h = h * 1000003L + (v + 1);
+			h = (long)((unsigned long)h * 1000003UL +
+		    (unsigned long)(v + 1));
 			atomic_store_explicit(&g_hash, h, memory_order_relaxed);
 		}
 	}
