+++ test/sim/test_sim_buggify2.c
@@ -65,7 +65,8 @@ leaf(xtc_task_t *self, void *u)
 	(void)self;
 	atomic_fetch_add_explicit(&g_done, 1, memory_order_relaxed);
 	h = atomic_load_explicit(&g_hash, memory_order_relaxed);
-	h = h * 1000003L + (id + 1);       /* ORDER-sensitive fold */
+	h = (long)((unsigned long)h * 1000003UL +
+		    (unsigned long)(id + 1));       /* ORDER-sensitive fold */
 	atomic_store_explicit(&g_hash, h, memory_order_relaxed);
 	return XTC_TASK_DONE;
 }
