+++ test/sim/test_sim_chan.c
@@ -81,7 +81,8 @@ static void
 mpmc_fold(int v)
 {
 	long h = atomic_load_explicit(&g_mpmc_hash, memory_order_relaxed);
-	h = h * 1000003L + (v + 1);
+	h = (long)((unsigned long)h * 1000003UL +
+		    (unsigned long)(v + 1));
 	atomic_store_explicit(&g_mpmc_hash, h, memory_order_relaxed);
 }
 
@@ -263,7 +264,8 @@ static void
 bc_fold(int v)
 {
 	long h = atomic_load_explicit(&g_bc_hash, memory_order_relaxed);
-	h = h * 1000003L + (v + 1);
+	h = (long)((unsigned long)h * 1000003UL +
+		    (unsigned long)(v + 1));
 	atomic_store_explicit(&g_bc_hash, h, memory_order_relaxed);
 }
