+++ test/sim/test_sim_lwlock.c
@@ -70,7 +70,8 @@ static void
 ex_fold(long v)
 {
 	long h = atomic_load_explicit(&g_ex_hash, memory_order_relaxed);
-	h = h * 1000003L + (v + 1);
+	h = (long)((unsigned long)h * 1000003UL +
+		    (unsigned long)(v + 1));
 	atomic_store_explicit(&g_ex_hash, h, memory_order_relaxed);
 }
 
@@ -213,7 +214,8 @@ rw_reader(void *arg)
 		{
 			long h = atomic_load_explicit(&g_rw_hash,
 			    memory_order_relaxed);
-			h = h * 1000003L + (hi + 1) + (long)id;
+			h = (long)((unsigned long)h * 1000003UL +
+			    (unsigned long)((hi + 1) + (long)id));
 			atomic_store_explicit(&g_rw_hash, h,
 			    memory_order_relaxed);
 		}
