+++ test/sim/test_sim_machine_death.c
@@ -62,7 +62,8 @@ static void
 fold(long v)
 {
 	long h = atomic_load_explicit(&g_evt_hash, memory_order_relaxed);
-	h = h * 1000003L + (v + 11);
+	h = (long)((unsigned long)h * 1000003UL +
+		    (unsigned long)(v + 11));
 	atomic_store_explicit(&g_evt_hash, h, memory_order_relaxed);
 }
